-- add function select to ibis::bin to work with clustered values
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@434 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/tab.h
@@ -33,7 +33,7 @@ class ibis::tabula : public ibis::table {
 public:
     tabula(const char* na, const char* de, uint64_t nr) :
 	table(na, de), nrows(nr) {};
-    explicit tabula(uint32_t nr=0) : nrows(nr) {};
+    explicit tabula(uint64_t nr=0) : nrows(nr) {};
     virtual ~tabula() {};
     virtual uint64_t nRows() const {return nrows;}
     virtual uint32_t nColumns() const {return 0;}
